web-app/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-login==0.6.2
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.7.0
pylint==2.17.4 